refactor(web-server): split messageHandlers into a directory module
- extract Auto Run and Tabs domains (23 handlers total) into autoRun.ts and tabs.ts
- add types.ts, shared.ts, and an index.ts barrel matching the MainPanel/ convention
- extracted handlers take an explicit MessageHandlerContext instead of `this`, so they don't depend on class state
- WebSocketMessageHandler.ts drops from 5,913 to 4,455 lines; pure code motion, no behavior change
