+++ src/main/web-server/handlers/messageHandlers/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Message Handlers Module Index
+ *
+ * Re-exports the WebSocket message handler class and its supporting types.
+ */
+
+export { WebSocketMessageHandler } from './WebSocketMessageHandler';
+export type {
+	WebClientMessage,
+	WebClient,
+	SessionDetailForHandler,
+	LiveSessionInfo,
+	MessageHandlerCallbacks,
+	MessageHandlerContext,
+} from './types';
