+++ src/main/web-server/handlers/messageHandlers/shared.ts
@@ -1,0 +1,5 @@
+/**
+ * Shared constants for WebSocket message handlers.
+ */
+
+export const LOG_CONTEXT = 'WebServer';
